backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-whois
pandas
scikit-learn
joblib
